backend: Fix activation bug and broken test…
During activation, the UserActivation was not being deleted because
the collection was not actually being retrieved and was therefore empty.
During troubleshooting, it became clear that the model was set up wrong.
The primary key was the default, meaning auto-incrementing 'id' column,
which doesn't exit. Change this to non-auto-incrementing 'token' column.